+++ Runtime/Core/SignalBroadcaster.cs
@@ -4,7 +4,6 @@ using UnityEngine;
 using NekoLib.Extensions;
 using NekoLib.Core;
 using NekoLib.Logger;
-using System.Diagnostics;
 
 #if UNITY_EDITOR
 using System.Linq;