+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.3.2] - 2025-11-08
+
+### Improve Signal Tracker and Monitor Editor
+
+- Add pagination support to Signal Receiver Monitor Editor.
+- Improve logging format in Signal Tracker Editor.
+
 ## [1.3.1] - 2025-11-06
 
 ### Rework Signal Tracker and Signal Receiver Monitor Editor